Add scripts to verify objects creation

+++ lib/pgdiff/models/schema.rb
@@ -14,11 +14,18 @@ module PgDiff
       end
 
       def add(diff)
+        return "" if nspname == "public"
         %Q{CREATE SCHEMA IF NOT EXISTS "#{nspname}";}
       end
 
       def remove(diff)
+        return "" if nspname == "public"
+
         %Q{DROP SCHEMA "#{nspname}";}
+      end
+
+      def change(diff, target)
+        return "" if nspname == "public"
       end
     end
   end
